[IMP] sale,stock,*: Convert day-based fields to integers
*= sale_stock,purchase_stock

Customer lead time defines the delay in days between order confirmation and
product shipment. Sales safety days add extra days to absorb supply chain
delays, replenishment horizon defines the number of days in advance
reordering rules are checked, and days to purchase defines the delay required
to confirm a purchase order. Fractional day values are not meaningful for
these fields.

This change enforces integer values for customer delivery days, sales safety
days, replenishment horizon, and the days to purchase, improving clarity and
consistency in sales and inventory planning while preserving the same visual
behavior as before.

Task - 4890380

+++ addons/purchase_stock/models/res_company.py
@@ -7,6 +7,6 @@ from odoo import fields, models
 class ResCompany(models.Model):
     _inherit = 'res.company'
 
-    days_to_purchase = fields.Float(
+    days_to_purchase = fields.Integer(
         string='Days to Purchase',
         help="Days needed to confirm a PO, define when a PO should be validated")
