+++ addons/purchase_stock/models/res_config_settings.py
@@ -8,7 +8,7 @@ class ResConfigSettings(models.TransientModel):
     _inherit = 'res.config.settings'
 
     module_stock_dropshipping = fields.Boolean("Dropshipping")
-    days_to_purchase = fields.Float(
+    days_to_purchase = fields.Integer(
         related='company_id.days_to_purchase', readonly=False)
     is_installed_sale = fields.Boolean(string="Is the Sale Module Installed")
